updated qtile fedora script

+++ qtile/install-qtile-fedora-wayland.sh
@@ -1,8 +1,13 @@
-qtile_src_dir="${HOME}/.local/src/qtile"
-qtile_repo_dir="${HOME}/.local/src/qtile/qtile"
+qtile_save_root="${HOME}/.local/src/qtile"
 link_qtile_bin_to="${HOME}/.local/bin"
 
-mkdir -p "${qtile_src_dir}"
+if [[ ! -d ${link_qtile_bin_to} ]]; then
+	mkdir -p ${link_qtile_bin_to}
+fi
+
+qtile_repo_dir="${qtile_save_root}/qtile"
+
+mkdir -p "${qtile_save_root}"
 
 if [[ ! -d "${qtile_repo_dir}" ]]; then
 	git clone https://github.com/qtile/qtile.git "${qtile_repo_dir}"
@@ -25,20 +30,23 @@ python_version="${major}.${minor}"
 echo "Detected Python version: $python_version"
 
 # Set up virtual environment
-if [[ ! -d "${qtile_src_dir}/.venv" ]]; then
-	python3 -m venv "${qtile_src_dir}/.venv"
+if [[ ! -d "${qtile_save_root}/.venv" ]]; then
+	python3 -m venv "${qtile_save_root}/.venv"
 fi
-source "${qtile_src_dir}/.venv/bin/activate"
-if [[ "$(which python)" != "${qtile_src_dir}/.venv/bin/python" ]]; then
+source "${qtile_save_root}/.venv/bin/activate"
+if [[ "$(which python)" != "${qtile_save_root}/.venv/bin/python" ]]; then
         echo "venv is not properly activated"
         echo "$(which python)"
         exit 1
 fi
 pip uninstall -y qtile
 
+# for v0.33.0
+# wlroots0.17 \
+# wlroots0.17-devel \
 sudo dnf install -y \
-	wlroots0.17 \
-	wlroots0.17-devel \
+	wlroots \
+	wlroots-devel \
 	wayland-protocols-devel \
 	libffi-devel \
 	cairo-devel \
@@ -54,18 +62,21 @@ pip install \
 
 cd "${qtile_repo_dir}"
 git fetch --tags
-git checkout v0.33.0
-pip install --no-cache-dir . --config-settings=backend=wayland qtile[wayland]
-# pip install . --config-settings=backend=wayland
+# git checkout v0.33.0
+# pip install --no-cache-dir . --config-settings=backend=wayland qtile[wayland]
+git checkout master
+pip install . --config-settings=backend=wayland
 
 # for non-fedora
 # pip install iwlib
 
-qtile_bin="${qtile_src_dir}/.venv/bin/qtile"
+qtile_bin="${qtile_save_root}/.venv/bin/qtile"
 if [[ ! -x "${qtile_bin}" ]]; then
-    echo "qtile bin not found in ${qtile_src_dir}/.venv/bin/"
+    echo "qtile bin not found in ${qtile_save_root}/.venv/bin/"
     exit 1
 fi
 
-unlink "${link_qtile_bin_to}/qtile"
+if [[ -f "${link_qtile_bin_to}/qtile" ]]; then
+	unlink "${link_qtile_bin_to}/qtile"
+fi
 ln -s "${qtile_bin}" "${link_qtile_bin_to}"
